IN_PROGRESS - bug 6277: BufferedDataTable should return number of rows as long https://bugs.knime.org/show_bug.cgi?id=6277
Adapted classes in org.knime.base

+++ org.knime.ext.textprocessing/src/org/knime/ext/textprocessing/nodes/frequencies/filter/KTermsFilter.java
@@ -41,11 +41,14 @@
  *  may freely choose the license terms applicable to such Node, including
  *  when such Node is propagated with or for interoperation with KNIME.
  * ---------------------------------------------------------------------
- * 
+ *
  * History
  *   22.04.2008 (thiel): created
  */
 package org.knime.ext.textprocessing.nodes.frequencies.filter;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import org.knime.base.data.sort.SortedTable;
 import org.knime.core.data.DataRow;
@@ -54,36 +57,34 @@ import org.knime.core.node.BufferedDataTable;
 import org.knime.core.node.CanceledExecutionException;
 import org.knime.core.node.ExecutionContext;
 
-import java.util.ArrayList;
-import java.util.List;
-
 /**
  * Keeps a specified number k of rows and filters out the rest. The k rows that
  * are kept are those with the highest value of a specified frequency.
- * 
+ *
  * @author Kilian Thiel, University of Konstanz
  */
 public class KTermsFilter extends FrequencyFilter {
-    
-    private int m_k;
-    
-    private int m_count = 0;
-    
+
+    private long m_k;
+
+    private long m_count = 0;
+
     /**
      * Creates a new instance of <code>KTermsFilter</code> with the given
-     * index of the term column, of the frequency column to apply the filtering 
+     * index of the term column, of the frequency column to apply the filtering
      * to and the number k of row to keep.
-     * 
+     *
      * @param termColIndex The index of the term column.
-     * @param colIndex The index of the frequency column to apply the filtering 
+     * @param colIndex The index of the frequency column to apply the filtering
      * to.
      * @param k The number k of rows to keep.
-     * @param modifyUnmodifiable if set <code>true</code>, unmodifiable terms 
-     * are modified or filtered even if they are set unmodifiable, otherwise 
+     * @param modifyUnmodifiable if set <code>true</code>, unmodifiable terms
+     * are modified or filtered even if they are set unmodifiable, otherwise
      * not.
+     * @since 3.0
      */
-    public KTermsFilter(final int termColIndex, final int colIndex, 
-            final int k, final boolean modifyUnmodifiable) {
+    public KTermsFilter(final int termColIndex, final int colIndex,
+            final long k, final boolean modifyUnmodifiable) {
         super(colIndex, termColIndex, modifyUnmodifiable);
         m_k = k;
     }
@@ -92,30 +93,30 @@ public class KTermsFilter extends FrequencyFilter {
      * {@inheritDoc}
      */
     @Override
-    public boolean internalMatches(final DataRow row, final int rowIndex) {
+    public boolean internalMatches(final DataRow row, final long rowIndex) {
         m_count++;
         if (m_count <= m_k) {
             return true;
         }
         return false;
     }
-    
+
     /**
      * {@inheritDoc}
-     * @throws CanceledExecutionException 
+     * @throws CanceledExecutionException
      */
     @Override
     public BufferedDataTable preprocessData(final BufferedDataTable data,
             final ExecutionContext exec) throws CanceledExecutionException {
         DataTableSpec spec = data.getDataTableSpec();
-        
+
         List<String> colName = new ArrayList<String>();
         colName.add(spec.getColumnSpec(m_filterColIndex).getName());
-        
+
         boolean[] sortAsc = new boolean[1];
         sortAsc[0] = false;
-        
+
         return new SortedTable(data, colName, sortAsc, exec).
                     getBufferedDataTable();
-    }   
+    }
 }
